Foreclosure read.me
This is code for my final project in Urban GIS at the University Of Minnesota. This code takes in data from a metro city that includes all foreclosures, when the sheriff sale took place, the address and how much the first bid was for the home after the foreclosure. I used Dplyr to clean the data so it could be joined with a parcel shapefile for the city. This shapefile can be found on the GIS open data website for Hennepin county. (http://www.hennepin.us/your-government/open-government/gis-open-data) This join was done so that the data could be mapped using ESRI's ArcGIS 10.4.